Tratamento de erros

+++ src/main/java/com/example/plantio/api/service/ZoneamentoService.java
@@ -1,7 +1,6 @@
 package com.example.plantio.api.service;
 
 import com.example.plantio.api.dto.ZoneamentoDTO;
-import com.example.plantio.api.exception.ObjectNotFoundException;
 import com.example.plantio.api.model.Zoneamento;
 import com.example.plantio.api.repository.ZoneamentoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,7 +23,7 @@ public class ZoneamentoService {
 
     public List<Object> getZoneamento() { return rep.findAll().stream().map(ZoneamentoDTO::create).collect(Collectors.toList()); };
 
-    public ZoneamentoDTO getZoneamentoById(Long id) { return rep.findById(id).map(ZoneamentoDTO::create).orElseThrow(() -> new ObjectNotFoundException("Zoneamento não encontreado"));}
+    public Optional<ZoneamentoDTO> getZoneamentoById(Long id) { return rep.findById(id).map(ZoneamentoDTO::create); }
 
     public List<ZoneamentoDTO> getZoneamentoBySafra(String safra) { return rep.findBySafra(safra).stream().map(ZoneamentoDTO::create).collect(Collectors.toList());}
 
@@ -65,8 +64,11 @@ public class ZoneamentoService {
         }
     }
 
-    public void delete(Long id) {
+    public boolean delete(Long id) {
+        if(getZoneamentoById(id).isPresent()) {
             rep.deleteById(id);
-
+            return true;
+        }
+        return false;
     }
 }
